refactor shadow configuration for spring boot
