+++ sharding-core/sharding-core-common/src/main/resources/META-INF/services/org.apache.shardingsphere.underlying.common.rule.ShardingSphereRuleBuilder
@@ -17,3 +17,4 @@
 
 org.apache.shardingsphere.core.rule.ShardingRuleBuilder
 org.apache.shardingsphere.core.rule.MasterSlaveRuleBuilder
+org.apache.shardingsphere.core.rule.ShadowRuleBuilder
