--- sharding-spring/sharding-jdbc-spring/sharding-jdbc-spring-boot-starter/src/test/resources/application-shadow-sharding.properties
@@ -1,96 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-spring.shardingsphere.datasource.names=ds,ds_${0..1},shadow_ds,shadow_ds_${0..1}
-spring.shardingsphere.datasource.ds.type=org.apache.commons.dbcp2.BasicDataSource
-spring.shardingsphere.datasource.ds.driver-class-name=org.h2.Driver
-spring.shardingsphere.datasource.ds.url=jdbc:h2:mem:ds;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MYSQL
-spring.shardingsphere.datasource.ds.username=sa
-spring.shardingsphere.datasource.ds.password=
-spring.shardingsphere.datasource.ds.max-total=100
-
-spring.shardingsphere.datasource.ds_0.type=org.apache.commons.dbcp2.BasicDataSource
-spring.shardingsphere.datasource.ds_0.driver-class-name=org.h2.Driver
-spring.shardingsphere.datasource.ds_0.url=jdbc:h2:mem:ds_0;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MYSQL
-spring.shardingsphere.datasource.ds_0.username=sa
-spring.shardingsphere.datasource.ds_0.password=
-spring.shardingsphere.datasource.ds_0.max-total=100
-
-spring.shardingsphere.datasource.ds_1.type=org.apache.commons.dbcp2.BasicDataSource
-spring.shardingsphere.datasource.ds_1.driver-class-name=org.h2.Driver
-spring.shardingsphere.datasource.ds_1.url=jdbc:h2:mem:ds_1;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MYSQL
-spring.shardingsphere.datasource.ds_1.username=sa
-spring.shardingsphere.datasource.ds_1.password=
-spring.shardingsphere.datasource.ds_1.max-total=100
-
-spring.shardingsphere.datasource.shadow_ds.type=org.apache.commons.dbcp2.BasicDataSource
-spring.shardingsphere.datasource.shadow_ds.driver-class-name=org.h2.Driver
-spring.shardingsphere.datasource.shadow_ds.url=jdbc:h2:mem:ds;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MYSQL
-spring.shardingsphere.datasource.shadow_ds.username=sa
-spring.shardingsphere.datasource.shadow_ds.password=
-spring.shardingsphere.datasource.shadow_ds.max-total=10
-
-spring.shardingsphere.datasource.shadow_ds_0.type=org.apache.commons.dbcp2.BasicDataSource
-spring.shardingsphere.datasource.shadow_ds_0.driver-class-name=org.h2.Driver
-spring.shardingsphere.datasource.shadow_ds_0.url=jdbc:h2:mem:ds_0;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MYSQL
-spring.shardingsphere.datasource.shadow_ds_0.username=sa
-spring.shardingsphere.datasource.shadow_ds_0.password=
-spring.shardingsphere.datasource.shadow_ds_0.max-total=10
-
-spring.shardingsphere.datasource.shadow_ds_1.type=org.apache.commons.dbcp2.BasicDataSource
-spring.shardingsphere.datasource.shadow_ds_1.driver-class-name=org.h2.Driver
-spring.shardingsphere.datasource.shadow_ds_1.url=jdbc:h2:mem:ds_1;DB_CLOSE_DELAY=-1;DATABASE_TO_UPPER=false;MODE=MYSQL
-spring.shardingsphere.datasource.shadow_ds_1.username=sa
-spring.shardingsphere.datasource.shadow_ds_1.password=
-spring.shardingsphere.datasource.shadow_ds_1.max-total=10
-
-        
-spring.shardingsphere.shadow.column=shadow
-spring.shardingsphere.shadow.shadow-mappings.ds=shadow_ds
-spring.shardingsphere.shadow.shadow-mappings.ds_0=shadow_ds_0
-spring.shardingsphere.shadow.shadow-mappings.ds_1=shadow_ds_1
-
-spring.shardingsphere.shadow.shardingRule.default-database-strategy.standard.sharding-column=user_id
-spring.shardingsphere.shadow.shardingRule.default-database-strategy.standard.sharding-algorithm.type=INLINE
-spring.shardingsphere.shadow.shardingRule.default-database-strategy.standard.sharding-algorithm.props.algorithm.expression=ds_${user_id % 2}
-
-spring.shardingsphere.shadow.shardingRule.tables.t_order.actual-data-nodes=ds_${0..1}.t_order_${0..1}
-spring.shardingsphere.shadow.shardingRule.tables.t_order.table-strategy.standard.sharding-column=order_id
-spring.shardingsphere.shadow.shardingRule.tables.t_order.table-strategy.standard.sharding-algorithm.type=INLINE
-spring.shardingsphere.shadow.shardingRule.tables.t_order.table-strategy.standard.sharding-algorithm.props.algorithm.expression=t_order_${order_id % 2}
-spring.shardingsphere.shadow.shardingRule.tables.t_order.key-generator.type=SNOWFLAKE
-spring.shardingsphere.shadow.shardingRule.tables.t_order.key-generator.column=order_id
-spring.shardingsphere.shadow.shardingRule.tables.t_order.key-generator.props.worker.id=123
-spring.shardingsphere.shadow.shardingRule.tables.t_order_item.actual-data-nodes=ds_${0..1}.t_order_item_${0..1}
-spring.shardingsphere.shadow.shardingRule.tables.t_order_item.table-strategy.standard.sharding-column=order_id
-spring.shardingsphere.shadow.shardingRule.tables.t_order_item.table-strategy.standard.sharding-algorithm.type=INLINE
-spring.shardingsphere.shadow.shardingRule.tables.t_order_item.table-strategy.standard.sharding-algorithm.props.algorithm.expression=t_order_item_${order_id % 2}
-spring.shardingsphere.shadow.shardingRule.tables.t_order_item.key-generator.type=SNOWFLAKE
-spring.shardingsphere.shadow.shardingRule.tables.t_order_item.key-generator.column=order_item_id
-spring.shardingsphere.shadow.shardingRule.tables.t_order_item.key-generator.props.worker.id=123
-spring.shardingsphere.shadow.shardingRule.binding-tables=t_order,t_order_item
-spring.shardingsphere.shadow.shardingRule.broadcast-tables=t_config
-spring.shardingsphere.shadow.shardingRule.encryptRule.encryptors.order_encrypt.type=test
-spring.shardingsphere.shadow.shardingRule.encryptRule.encryptors.order_encrypt.props.appToken=business
-spring.shardingsphere.shadow.shardingRule.encryptRule.tables.t_order.columns.pwd.cipherColumn=cipher_pwd
-spring.shardingsphere.shadow.shardingRule.encryptRule.tables.t_order.columns.pwd.encryptor=order_encrypt
-spring.shardingsphere.shadow.shardingRule.encryptRule.tables.t_order.columns.name.cipherColumn=cipher_name
-spring.shardingsphere.shadow.shardingRule.encryptRule.tables.t_order.columns.name.plainColumn=plain_name
-spring.shardingsphere.shadow.shardingRule.encryptRule.tables.t_order.columns.name.encryptor=order_encrypt
-
-spring.shardingsphere.props.sql.show=true
-spring.shardingsphere.props.executor.size=100
