+++ sharding-spring/sharding-jdbc-spring/sharding-jdbc-spring-boot-starter/src/test/resources/application-shadow.properties
@@ -31,8 +31,8 @@ spring.shardingsphere.datasource.shadow_ds.username=sa
 spring.shardingsphere.datasource.shadow_ds.password=
 spring.shardingsphere.datasource.shadow_ds.max-total=99
 
-spring.shardingsphere.shadow.column=is_shadow
-spring.shardingsphere.shadow.shadow-mappings.ds=shadow_ds
+spring.shardingsphere.rules.shadow-rule.column=is_shadow
+spring.shardingsphere.rules.shadow-rule.shadow-mappings.ds=shadow_ds
 
 spring.shardingsphere.props.sql.show=true
 spring.shardingsphere.props.executor.size=100
